requirements.txt
yfinance>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
requests>=2.32.0
feedparser>=6.0.0
rich>=13.0.0
typer>=0.12.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
